requirements.txt
torch
pandas
wikipedia-api
requests
python-dotenv
sentence-transformers